requirements.txt
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.6
cloud-init==25.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==46.0.3
cssselect2==0.8.0
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20241021
distro==1.9.0
distro-info==1.7+build1
dj-database-url==3.0.1
Django==5.2.7
django-allauth==65.13.1
django-axes==8.1.0
django-environ==0.12.0
django-heroku==0.3.1
django-import-export==4.3.12
django-simple-history==3.11.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
fonttools==4.60.1
freetype-py==2.5.1
gunicorn==23.0.0
gyp==0.1
html5lib==1.1
httplib2==0.20.4
idna==3.6
import-export==0.3.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
meson==1.3.2
mysqlclient==2.2.7
netaddr==0.8.0
netifaces==0.11.0
numpy==2.3.4
oauthlib==3.3.1
olefile==0.46
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.0
pandas==2.3.3
pexpect==4.9.0
pillow==10.2.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pwdsphinx==1.0.16
pycairo==1.25.1
pycparser==2.23
pycups==2.0.1
pydyf==0.11.0
pyequihash==0.2
Pygments==2.17.2
PyGObject==3.48.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyparsing==3.1.1
pypdf==6.1.3
pyphen==0.17.2
pyrsistent==0.20.0
pyserial==3.5
pysodium==0.7.0.post0
python-apt==2.7.7+ubuntu5.1
python-bidi==0.6.7
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-docx==1.2.0
python3-openid==3.2.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
qrcode==8.2
qrcodegen==1.8.0
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.7.1
rlPyCairo==0.4.0
SecureString==0.2
setuptools==80.9.0
six==1.16.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
svglib==1.6.0
systemd-python==235
tablib==3.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.10.0
tzdata==2025.2
tzlocal==5.3.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritools==5.0.0
urllib3==2.0.7
wadllib==1.3.6
weasyprint==66.0
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.11.0
xdg==5
xhtml2pdf==0.2.17
xkit==0.0.0
zopfli==0.2.3.post1
zxcvbn==4.4.28
